refactor: Refactoring the image upload process and generating ShortCodes
Each processed image is treated as a transaction using UnitOfWork. The construction of a ShortednedUrl has been moved to inside UploadImage, which now returns the shortCodes of each processed image. The ShortCodeGeneratorService interface has been removed and replaced by ShortCodeGenerator, which is responsible only for generating and verifying a new shortCode.